Adding getAllFocusableDescendants to helpers. Added tests.
